pyproject.toml
[project]
name = "my-garmin-api"
version = "0.1.0"
description = "Personal Garmin API"
readme = "README.md"
authors = [
    {name = "Adam Jakab", email = "adam@jakab.pro"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "curl-cffi (>=0.15.0,<0.16.0)",
    "fastapi (>=0.100.0,<1.0.0)",
    "garminconnect (>=0.3.2,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "ua-generator (>=2.0.24,<3.0.0)",
    "uvicorn (>=0.24.0,<1.0.0)",
]

[project.scripts]
api = "my_garmin_api.api:main"
api-dev = "my_garmin_api.api:main_dev"


[tool.pylint.MASTER]
ignore-paths = ["my_garmin_api/garminconnect"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
